pyproject.toml
[tool.poetry]
name = "bd_metamusic"
version = "0.1.0"
description = ""
authors = ["Magno"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1"
djangorestframework = "^3.13.1"
django-cors-headers = "^3.13.0"
gunicorn = "^20.1.0"
whitenoise = "^6.2.0"
djangorestframework-simplejwt = "^5.2.0"
python-magic = "^0.4.27"
Pillow = "^9.2.0"
drf-spectacular = "^0.24.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
